backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
requests==2.32.3
python-dotenv==1.0.1
chromadb>=1.0.0
sentence-transformers==3.3.1
openai==1.54.0
python-multipart==0.0.12
